feat: menu do usuário na página principal
